Move ReadLayeredYAMLs to mesh/shared.go
