+++ cmd/mesh/shared.go
@@ -19,8 +19,11 @@ package mesh
 import (
 	"fmt"
 	"io"
+	"io/ioutil"
 	"os"
+	"strings"
 
+	"istio.io/operator/pkg/util"
 	"istio.io/pkg/log"
 )
 
@@ -121,4 +124,19 @@ func (l *Logger) printErr(s string) {
 
 func refreshGoldenFiles() bool {
 	return os.Getenv("REFRESH_GOLDEN") == "true"
+}
+
+func ReadLayeredYAMLs(filenames []string) (string, error) {
+	var ly string
+	for _, fn := range filenames {
+		b, err := ioutil.ReadFile(strings.TrimSpace(fn))
+		if err != nil {
+			return "", err
+		}
+		ly, err = util.OverlayYAML(ly, string(b))
+		if err != nil {
+			return "", err
+		}
+	}
+	return ly, nil
 }
