+++ cmd/mesh/shared_test.go
@@ -1,38 +1,12 @@
-// Copyright 2019 Istio Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package util
+package mesh
 
 import (
 	"io/ioutil"
-	"os"
 	"path/filepath"
 	"testing"
-)
 
-var (
-	repoRootDir string
-	testDataDir string
+	"istio.io/operator/pkg/util"
 )
-
-func init() {
-	wd, err := os.Getwd()
-	if err != nil {
-		panic(err)
-	}
-	repoRootDir = filepath.Join(wd, "../..")
-}
 
 func TestReadLayeredYAMLs(t *testing.T) {
 	testDataDir = filepath.Join(repoRootDir, "pkg/util/testdata/yaml")
@@ -77,7 +51,7 @@ func TestReadLayeredYAMLs(t *testing.T) {
 				return
 			}
 
-			if YAMLDiff(got, want) != "" {
+			if util.YAMLDiff(got, want) != "" {
 				t.Errorf("ReadLayeredYAMLs() got = %v, want %v", got, want)
 			}
 		})
