+++ pkg/util/yaml.go
@@ -17,7 +17,6 @@ package util
 import (
 	"bytes"
 	"fmt"
-	"io/ioutil"
 	"strings"
 
 	jsonpatch "github.com/evanphx/json-patch"
@@ -166,21 +165,6 @@ func OverlayYAML(base, overlay string) (string, error) {
 	}
 
 	return string(my), nil
-}
-
-func ReadLayeredYAMLs(filenames []string) (string, error) {
-	var ly string
-	for _, fn := range filenames {
-		b, err := ioutil.ReadFile(strings.TrimSpace(fn))
-		if err != nil {
-			return "", err
-		}
-		ly, err = OverlayYAML(ly, string(b))
-		if err != nil {
-			return "", err
-		}
-	}
-	return ly, nil
 }
 
 func YAMLDiff(a, b string) string {
